Shopping Assistant feature
Add new page for shopping assistant with form. Here a user a enter text for exploring clothing suggestions from ChatGPT based on the text and the database of products. If the search param called 'query' is used in the URL, the serach is started as soon as the page load. This is so that searches from the home page is redirected here and the serach starts. Otherwise, a user can search as usual through the form. On the backend, some prompt engineering has been tried to improve the results from OpenAI's API, including cases where no products could be suggested resulting in returning an empty array. Each user search text is now also saved in the database for later use.