Improve Product page
Improved the design of the product page. A lot of effort has gone into getting the responsiveness right and good looking. Add some customer benefits icons. Add related products section on the page. For this a new API route was created which fetch similar products based on the type and gender of the product in focus.