Add meta data
Add page titles to most pages on the website. Dynamic page title for product page with brand name. Add Open Graph meta data for social media linking to website but only for home page. For this an API endpoint was added for website logo retrieval.

Small bug fixes for related products on the product page where tailwind were not able to dynamically set the number of grid cols based on the number of related products. Lastly, a small misstake corrected for svg attributes like stroke-width to work with React.