Improve the nav menu for mobile (smaller screens).
A new library Framer Motion was added to get a nice looking animation for opening and closing the menu. Some smaller changes was also made like hovering color for the links in the menu.

A small adjustment was made to the padding on the page for the shopping assistant noticed during testing on in a smaller window which now should look better.