Fix infinite product card loading icon.
Instead now a default product image not found is shown uploaded with this commit. Since I discovered that i had three different ProductCard components, i decided to refactor my code and make only one as it should be. Added some components props to handle the difference in style which was the only difference between the original three ProductCard components. Remove the two local ProductCard components. With this I also had to modify one test to work with the new ProductCard. Small change of alt text for image on the product page.