Completed Contact form and page
Completed the backend for the contact form by saving the input from the form to the MongoDB in a new collection. For this a new mongoose model was added and code to create new documents and fetch them as well. In the dashboard, a new page was created to look and read at the contact requests received.

Changes some About Brilliantwear text.

Update README.

On the fron-end side, a message was added to be shown when the form has been submitted. Lastly, disabled the submit button after a user clicks on it to not accidentally send the form more than once per page load and fill.