Add Privacy Notice page for GDPR
A new page was added for information regarding the privacy notice of user's data when using Brilliantwear.

A disclaimer was add to both the Terms and Condition & Privacy Notice pages to make it clear that not all information on these pages should be taken too serious.

Link added to new Privacy Notice page has been added to the footer.