Add Cookie banner for consent.
A cookie banner was added to the bottom of the viewport. It was added to the root layout.tsx to show up on every page until accepted. Some animation was added between the closed and open state with duration. A lot of effort went into getting the colors right in the banner to look good with the rest of the website while still standing out from the rest to catch the user's attention.

Some information about how cookies are used was added to the Privacy Notice page. The list of personal information we collect was changed by having each address information in its own bullet instead of within parentheses placed along side Address.

Update README.md