Remove the possibility for anyone to register for the Admin dashboard.
Register API endpoint removed from the public accessible paths. Public register page with layout removed. Link to this page was also removed on the login page.

New dashboard navigation bar added with links some of the pages available for the admin. Also, previously missing log out button added to this nav bar. This nav bar is only shown when logged in as an admin at the top below the common nav bar across the website.

Remove the not working returnUrl functionality when a someone tries to access any of the dashboard pages without being logged in which should have returned to that page when a user has logged in. This did not work and always when an admin user logged in, it directed the admin to the home page of Brilliantwear. Now instead, the user is always directed to the dashboard home page after a successful login.