Prepare for deploying verion 1.0.0 of Brilliantwear
Modify the code to use &quot; and &apos; in order to compile it.

Fix error handling on home page if the API can't be reached to avoid error.

Fix bug where generateMetadata was used in a layout.tsx file. Instead this layout file was removed and the page file was made a server component that imports an entire page component for the product.

Set version to 1.0.0 in package.json