Fix front-end bug where the entire page could be scrolled horizontally. Add some fail control when a user enters an invalid address.
Now all of the checkout pages should not horizontally scroll on smaller screens. Added x-padding to make the content look better and not go all the way to the edges.

Add some x-padding to other pages like the shopping assistant to look better on smaller screens.

On the shipping information checkout page, now when a user enter an incorrect address which the PostNord API can't fetch, a messeage is shown for the user to check the entered address. Previously, an error would occur but not anymore.